IGNITE-24333 Extract Node class from ItReplicaLifecycleTest
